Fixed long conversion

+++ logicaldoc-gui/src/main/java/com/logicaldoc/gui/frontend/client/tenant/TenantQuotaPanel.java
@@ -2,6 +2,8 @@ package com.logicaldoc.gui.frontend.client.tenant;
 
 import java.util.Map;
 
+import org.apache.bcel.generic.NEW;
+
 import com.logicaldoc.gui.common.client.beans.GUITenant;
 import com.logicaldoc.gui.common.client.data.UsersDS;
 import com.logicaldoc.gui.common.client.util.ItemFactory;
@@ -84,7 +86,7 @@ public class TenantQuotaPanel extends HLayout {
 		SpinnerItem usersQuota = ItemFactory.newSpinnerItem(USERSQUOTA, tenant.getMaxUsers());
 		usersQuota.setDisabled(readonly);
 		usersQuota.setRequired(false);
-		usersQuota.setMin(tenant.getUsers());
+		usersQuota.setMin((double)tenant.getUsers());
 		usersQuota.setStep(1);
 		usersQuota.setWidth(80);
 		if (!readonly)
@@ -93,7 +95,7 @@ public class TenantQuotaPanel extends HLayout {
 		SpinnerItem guestsQuota = ItemFactory.newSpinnerItem(GUESTSQUOTA, "readonlyusersquota", tenant.getMaxGuests());
 		guestsQuota.setDisabled(readonly);
 		guestsQuota.setRequired(false);
-		guestsQuota.setMin(tenant.getGuests());
+		guestsQuota.setMin((double)tenant.getGuests());
 		guestsQuota.setStep(1);
 		guestsQuota.setWidth(80);
 		if (!readonly)
